backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.8.2,<3.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
graphiti-core==0.3.0
openai>=1.38.0,<2.0.0
numpy==1.24.3
pandas==2.0.3
spacy==3.7.2
networkx==3.2.1
plotly==5.17.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
supabase==2.3.4
asyncpg==0.29.0
websockets==12.0
